Backend/Microservices/like_service/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
httpx
tenacity
python-dotenv

